+++ CHANGES.md
@@ -1,5 +1,10 @@
 # **libCLImate.Go** Changes
 
+## 0.3.0 - 28th March 2019
+
+* + added ``Climate.ProgramName`` field
+* + added ``Climate.Abort()`` method
+
 ## 0.2.0 - 25th March 2019
 
 * + added ``Climate.AddFlagFunc()`` and ``Climate.AddOptionFunc()`` methods, which take callbacks to be executed when the flag/option is encountered in the parsed command-line arguments;